[arcane,alina] Continue transform 'unsigned' type to 'signed' types.

+++ arcane/src/arcane/alina/ParmetisMatrixPartitioner.h
@@ -51,7 +51,7 @@ struct ParmetisMatrixPartitioner
   struct params
   {
     bool shrink;
-    ptrdiff_t min_per_proc;
+    int min_per_proc;
     int shrink_ratio;
 
     params()
